Former-commit-id: 023767a81382337e53e941d7e3bdf590cd7c7e8e

+++ sql/tool/list-operational-scripts.sh
@@ -23,6 +23,15 @@ function print_banner() {
     echo "******************************************************************************************"
 }
 
+function registerSQLScript() {
+    git log $1 >> /dev/null 2>&1
+    if [ $? -eq 0 ]; then
+        HASH=`git rev-list HEAD $1 | tail -n 1`
+        DATE=`git show -s --format="%ct" $HASH --`
+        printf "%s#%s \n" "$DATE" $1 >> $2
+    fi
+}
+
 clear
 
 if [ "$0" != "./sql/tool/$(basename $0)" ]; then
@@ -38,7 +47,7 @@ fi
 if [ "$#" -lt 1 ]; then
     print_banner
     echo ""
-    echo "   Uso: $0 [haya|bankia]"
+    echo "   Uso: $0 [cliente]"
     echo ""
     echo ""
     echo "******************************************************************************************"
@@ -49,23 +58,44 @@ fi
 print_banner
 
 CUSTOMER_IN_LOWERCASE=`echo $1 | tr '[:upper:]' '[:lower:]'`
-
+CUSTOMER_IN_UPPERCASE=`echo $1 | tr '[:lower:]' '[:upper:]'`
 BASEDIR=$(dirname $0)
+
+export SETENVGLOBAL=~/setEnvGlobal.sh
+if [ -f ~/setEnvGlobal${CUSTOMER_IN_UPPERCASE}.sh ] ; then
+  export SETENVGLOBAL=~/setEnvGlobal${CUSTOMER_IN_UPPERCASE}.sh
+fi
+if [ ! -f $SETENVGLOBAL ]; then
+    echo "No existe el fichero: $SETENVGLOBAL"
+    echo "Consulta las plantillas que hay en sql/tool/templates"
+    exit 1
+fi
+source $SETENVGLOBAL
 
 rm -rf $BASEDIR/tmp/*.txt
 
-for directory in `find ./sql -mindepth 1 -maxdepth 1 -name '?\.*'`
+for file in `ls sql/**/producto/*.sql`
 do
-    for file in `find $directory -maxdepth 4 -type f -name *.sql`
-    do
-        git log $file >> /dev/null 2>&1
-        if [ $? -eq 0 ]; then
-            HASH=`git rev-list HEAD $file | tail -n 1`    
-            DATE=`git show -s --format="%ct" $HASH --`   
-            printf "%s#%s \n" "$DATE" $file >> $BASEDIR/tmp/from-date-list-1.txt
-        fi
-    done
+    registerSQLScript $file $BASEDIR/tmp/product-list.txt
 done
 
-cat $BASEDIR/tmp/from-date-list-1.txt | grep "producto" | sort | cut -d# -f2 
-cat $BASEDIR/tmp/from-date-list-1.txt | grep "$CUSTOMER_IN_LOWERCASE" | sort | cut -d# -f2 
+for file in `ls sql/**/$CUSTOMER_IN_LOWERCASE/*.sql`
+do
+    registerSQLScript $file $BASEDIR/tmp/customer-list.txt
+done
+
+cat $BASEDIR/tmp/product-list.txt | sort | cut -d# -f2
+
+if [ "$MULTIENTIDAD" != "" ] ; then
+    IFS=',' read -a entidades <<< "$MULTIENTIDAD"
+    for entidad in "${entidades[@]}"
+    do
+        SUBENTITY=`echo $entidad | tr '[:upper:]' '[:lower:]'`
+        for file in `ls sql/**/$CUSTOMER_IN_LOWERCASE/$SUBENTITY/*.sql`
+        do
+            registerSQLScript $file $BASEDIR/tmp/customer-list.txt
+        done
+    done
+fi
+
+cat $BASEDIR/tmp/customer-list.txt | sort | cut -d# -f2
